requirements.txt
certifi==2025.6.15
charset-normalizer==3.4.2
colorama==0.4.6
filelock==3.18.0
fsspec==2025.5.1
idna==3.10
Jinja2==3.1.6
joblib==1.5.1
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.5
numpy==2.3.1
pillow==11.3.0
python-dotenv==1.1.0
requests==2.32.4
scikit-learn==1.7.0
scipy==1.16.0
setuptools==80.9.0
sympy==1.14.0
threadpoolctl==3.6.0
torch==2.7.1
torchvision==0.22.1
tqdm==4.67.1
typing_extensions==4.14.0
urllib3==2.5.0
